+++ mcrit/Worker.py
@@ -501,9 +501,17 @@ class Worker(QueueRemoteCallee):
 
     # Reports PROGRESS
     @Remote(progress=True, json_locations=[0])
-    def getMatchesForSmdaReport(self, report_json, minhash_threshold=None, pichash_size=None, band_matches_required=None, progress_reporter=NoProgressReporter()):
+    def getMatchesForSmdaReport(
+        self, report_json, minhash_threshold=None, pichash_size=None, band_matches_required=None, shortlist_size=None, band_df_cutoff=None, progress_reporter=NoProgressReporter()
+    ):
         matcher = MatcherQuery(
-            self, minhash_threshold=minhash_threshold, pichash_size=pichash_size, band_matches_required=band_matches_required, progress_reporter=progress_reporter
+            self,
+            minhash_threshold=minhash_threshold,
+            pichash_size=pichash_size,
+            band_matches_required=band_matches_required,
+            progress_reporter=progress_reporter,
+            shortlist_size=shortlist_size,
+            band_df_cutoff=band_df_cutoff,
         )
         smda_report = SmdaReport.fromDict(report_json)
         match_report = matcher.getMatchesForSmdaReport(smda_report)
@@ -511,35 +519,67 @@ class Worker(QueueRemoteCallee):
 
     # Reports PROGRESS
     @Remote(progress=True, file_locations=[0])
-    def getMatchesForMappedBinary(self, binary, base_address, minhash_threshold=None, pichash_size=None, band_matches_required=None, progress_reporter=NoProgressReporter()):
+    def getMatchesForMappedBinary(
+        self,
+        binary,
+        base_address,
+        minhash_threshold=None,
+        pichash_size=None,
+        band_matches_required=None,
+        shortlist_size=None,
+        band_df_cutoff=None,
+        progress_reporter=NoProgressReporter(),
+    ):
         config = SmdaConfig()
         SMDA_REPORT = None
         DISASSEMBLER = Disassembler(config)
         SMDA_REPORT = DISASSEMBLER.disassembleBuffer(binary, base_address)
         matcher = MatcherQuery(
-            self, minhash_threshold=minhash_threshold, pichash_size=pichash_size, band_matches_required=band_matches_required, progress_reporter=progress_reporter
+            self,
+            minhash_threshold=minhash_threshold,
+            pichash_size=pichash_size,
+            band_matches_required=band_matches_required,
+            progress_reporter=progress_reporter,
+            shortlist_size=shortlist_size,
+            band_df_cutoff=band_df_cutoff,
         )
         match_report = matcher.getMatchesForSmdaReport(SMDA_REPORT)
         return match_report
 
     # Reports PROGRESS
     @Remote(progress=True, file_locations=[0])
-    def getMatchesForUnmappedBinary(self, binary, minhash_threshold=None, pichash_size=None, band_matches_required=None, progress_reporter=NoProgressReporter()):
+    def getMatchesForUnmappedBinary(
+        self, binary, minhash_threshold=None, pichash_size=None, band_matches_required=None, shortlist_size=None, band_df_cutoff=None, progress_reporter=NoProgressReporter()
+    ):
         config = SmdaConfig()
         SMDA_REPORT = None
         DISASSEMBLER = Disassembler(config)
         SMDA_REPORT = DISASSEMBLER.disassembleUnmappedBuffer(binary)
         matcher = MatcherQuery(
-            self, minhash_threshold=minhash_threshold, pichash_size=pichash_size, band_matches_required=band_matches_required, progress_reporter=progress_reporter
+            self,
+            minhash_threshold=minhash_threshold,
+            pichash_size=pichash_size,
+            band_matches_required=band_matches_required,
+            progress_reporter=progress_reporter,
+            shortlist_size=shortlist_size,
+            band_df_cutoff=band_df_cutoff,
         )
         match_report = matcher.getMatchesForSmdaReport(SMDA_REPORT)
         return match_report
 
     # Reports PROGRESS
     @Remote(progress=True)
-    def getMatchesForSample(self, sample_id, minhash_threshold=None, pichash_size=None, band_matches_required=None, progress_reporter=NoProgressReporter()):
+    def getMatchesForSample(
+        self, sample_id, minhash_threshold=None, pichash_size=None, band_matches_required=None, shortlist_size=None, band_df_cutoff=None, progress_reporter=NoProgressReporter()
+    ):
         matcher = MatcherSample(
-            self, minhash_threshold=minhash_threshold, pichash_size=pichash_size, band_matches_required=band_matches_required, progress_reporter=progress_reporter
+            self,
+            minhash_threshold=minhash_threshold,
+            pichash_size=pichash_size,
+            band_matches_required=band_matches_required,
+            progress_reporter=progress_reporter,
+            shortlist_size=shortlist_size,
+            band_df_cutoff=band_df_cutoff,
         )
         match_report = matcher.getMatchesForSample(sample_id)
         return match_report
@@ -553,19 +593,39 @@ class Worker(QueueRemoteCallee):
         minhash_threshold=None,
         pichash_size=None,
         band_matches_required=None,
+        band_df_cutoff=None,
         progress_reporter=NoProgressReporter(),
     ):
-        matcher = MatcherVs(self, minhash_threshold=minhash_threshold, pichash_size=pichash_size, band_matches_required=band_matches_required, progress_reporter=progress_reporter)
+        matcher = MatcherVs(
+            self,
+            minhash_threshold=minhash_threshold,
+            pichash_size=pichash_size,
+            band_matches_required=band_matches_required,
+            progress_reporter=progress_reporter,
+            band_df_cutoff=band_df_cutoff,
+        )
         match_report = matcher.getMatchesForSample(sample_id, other_sample_id)
         return match_report
 
     # Reports PROGRESS
     @Remote(progress=True)
     def getMatchesForSampleVsGroup(
-        self, sample_id, other_sample_ids: List[int], minhash_threshold=None, pichash_size=None, band_matches_required=None, progress_reporter=NoProgressReporter()
+        self,
+        sample_id,
+        other_sample_ids: List[int],
+        minhash_threshold=None,
+        pichash_size=None,
+        band_matches_required=None,
+        band_df_cutoff=None,
+        progress_reporter=NoProgressReporter(),
     ):
         matcher = MatcherVsGroup(
-            self, minhash_threshold=minhash_threshold, pichash_size=pichash_size, band_matches_required=band_matches_required, progress_reporter=progress_reporter
+            self,
+            minhash_threshold=minhash_threshold,
+            pichash_size=pichash_size,
+            band_matches_required=band_matches_required,
+            progress_reporter=progress_reporter,
+            band_df_cutoff=band_df_cutoff,
         )
         match_report = matcher.getMatchesForSample(sample_id, other_sample_ids)
         return match_report